+++ packages/cli/src/commands/generate.ts
@@ -65,6 +65,26 @@ export async function generate(opts: GenerateOptions) {
     })),
   )
 
+  const contracts = Object.fromEntries(
+    await Promise.all(
+      (["ink", "sol"] as const).map(async (kind) => [
+        kind,
+        Object.fromEntries(
+          await Promise.all(
+            Object.entries(config[kind] ?? {}).map(async ([key, v]) => [
+              key,
+              toHex(
+                Blake2128(
+                  Binary.fromText(await fs.readFile(v, "utf-8")).asBytes(),
+                ),
+              ),
+            ]),
+          ),
+        ),
+      ]),
+    ),
+  )
+
   if (chains.length === 0) {
     console.log("No entries in polkadot-api.json. Nothing to generate.")
     return
@@ -74,7 +94,7 @@ export async function generate(opts: GenerateOptions) {
   const whitelist = whitelistPath ? await readWhitelist(whitelistPath) : null
 
   const descriptorsDir = join(process.cwd(), config.descriptorPath)
-  if (await alreadyGenerated(descriptorsDir, chains, whitelist)) {
+  if (await alreadyGenerated(descriptorsDir, chains, contracts, whitelist)) {
     console.log("Detected previous descriptors with no changes needed.")
     return
   }
@@ -102,7 +122,7 @@ export async function generate(opts: GenerateOptions) {
   await replacePackageJson(descriptorsDir, hash)
   const cleanCodegen = await compileCodegen(descriptorsDir)
   if (cleanCodegen) {
-    await tagGenerated(descriptorsDir, chains, whitelist)
+    await tagGenerated(descriptorsDir, chains, contracts, whitelist)
   }
   await fs.rm(descriptorSrcDir, { recursive: true })
   if (!config.options?.noDescriptorsPackage) {
@@ -117,6 +137,7 @@ async function tagGenerated(
     key: string
     metadataRaw: Uint8Array
   }[],
+  contracts: Record<"ink" | "sol", Record<string, HexString>>,
   whitelist: string[] | null,
 ) {
   const filePath = join(descriptorsDir, "generated.json")
@@ -132,6 +153,7 @@ async function tagGenerated(
           toHex(Blake2128(metadataRaw)),
         ]),
       ),
+      contracts,
     }),
   )
 }
@@ -141,6 +163,7 @@ async function alreadyGenerated(
     key: string
     metadataRaw: Uint8Array
   }[],
+  contracts: Record<"ink" | "sol", Record<string, HexString>>,
   whitelist: string[] | null,
 ) {
   const generatedJsôn = join(descriptorsDir, "generated.json")
@@ -164,7 +187,15 @@ async function alreadyGenerated(
       chains.every(({ key, metadataRaw }) => {
         const hash = toHex(Blake2128(metadataRaw))
         return hash === generated.chains[key]
-      })
+      }) &&
+      (["ink", "sol"] as const).every(
+        (kind) =>
+          Object.entries(contracts[kind]).length ===
+            Object.entries(generated.contracts[kind]).length &&
+          Object.entries(contracts[kind]).every(
+            ([k, hash]) => hash === generated.contracts[kind][k],
+          ),
+      )
     )
   } catch {
     return false